% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_criteria.r
\name{determine_number_of_dynamic_factors}
\alias{determine_number_of_dynamic_factors}
\title{Determine Number of Dynamic Factors}
\usage{
determine_number_of_dynamic_factors(data, method = "av", max_factors = 10)
}
\arguments{
\item{data}{A matrix of data to be analyzed.}

\item{method}{Method to determine factors. Must be one of "on", "av", or "hl". Default is "av".}

\item{max_factors}{Maximum number of factors. Default is 10.}
}
\value{
The number of dynamic factors.
}
\description{
This function determines the number of dynamic factors in the data using specified methods.
}
\details{
This function uses the following methods to determine the number of dynamic factors:
\itemize{
  \item \code{"on"}: Onatski (2009) test.
  \item \code{"av"}: Avarucci et al. (2022) test - relies on the `fnets` package.
  \item \code{"hl"}: Hallin & Liska (2007) test - relies on the `fnets` package.
}
}
\references{
Onatski, A. (2009). Testing hypotheses about the number of factors in large factor models. Econometrica, 77(5), 1447–1479.
Avarucci, M., Cavicchioli, M., Mario, F., & Zaffaroni, P. (2021, 01). The main business cycle shock(s). frequency-band estimation of the number of dynamic factors.
Hallin, M., & Liska, R. (2007, 02). Determining the number of factors in the general dynamic factor model. Journal of the American Statistical Association, 102,603-617.
}
