% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_criteria.r
\name{determine_number_of_static_factors}
\alias{determine_number_of_static_factors}
\title{Determine Number of Static Factors}
\usage{
determine_number_of_static_factors(data, method = "on", max_factors = 10)
}
\arguments{
\item{data}{A matrix of data to be analyzed.}

\item{method}{Method to determine factors. Must be one of "on", "al", "ah", or "bn". Default is "on".}

\item{max_factors}{Maximum number of factors. Default is 10.}
}
\value{
The number of static factors.
}
\description{
This function determines the number of static factors in the data using specified methods.
}
\details{
This function uses the following methods to determine the number of static factors:
\itemize{
  \item \code{"on"}: Onatski (2010) test.
  \item \code{"al"}: Alessi, Barigozzi and Capasso (2010) test - relies on the `fnets` package.
  \item \code{"ah"}: Ahn and Horenstein (2013) test - relies on the `fnets` package.
  \item \code{"bn"}: Bai and Ng (2002) test - relies on the `dfms` package.
}
}
\references{
Onatski, A. (2010). Determining the number of factors from empirical distribution of eigenvalues. The Review of Economics and Statistics, 92(4), 1004–1016.
Alessi, L., Barigozzi, M., & Capasso, M. (2010, December 1). Improved penalization for determining the number of factors in approximate factor models. Statistics Probability Letters, 80(23-24), 1806–1813
Ahn, S., & Horenstein, A. (2013, May 1). Eigenvalue ratio test for the number of factors. Econometrica, 81(3), 1203–1227
Bai, J., & Ng, S. (2003). Determining the number of factors in approximate factor models.
}
