% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_criteria.r
\name{dfms_ic_wrapper}
\alias{dfms_ic_wrapper}
\title{Wrap DFMS IC Results}
\usage{
dfms_ic_wrapper(dfms_ic)
}
\arguments{
\item{dfms_ic}{A list containing the DFMS IC results, specifically the "r.star" element.}
}
\value{
A data frame with the number of factors determined by the DFMS IC method.
}
\description{
This function wraps the results from the DFMS IC method into a data frame.
}
\references{
Bai, J., & Ng, S. (2002). Determining the Number of Factors in Approximate Factor Models. \emph{Econometrica}, 70(1), 191-221.
}
