% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm.R
\name{estimate_DFM}
\alias{estimate_DFM}
\title{Estimate Dynamic Factor Model (DFM)}
\usage{
estimate_DFM(
  data,
  r = NULL,
  q = NULL,
  q_max = 10,
  M = NULL,
  model = "fhlr_2000"
)
}
\arguments{
\item{data}{T x N dimensional matrix of data to be analyzed. Data can be in the form of a matrix or data frame.}

\item{r}{Number of static factors in the model. If NULL, it will be determined automatically.}

\item{q}{Number of dynamic factors in the model. If NULL, it will be determined automatically.}

\item{q_max}{Maximum number of dynamic factors to consider. Default is 10.}

\item{M}{Number of lags for the spectral density matrix. If NULL, it will be calculated as (2/3) * length(data[,1])^(1/3). Default is NULL.}

\item{model}{The model to use for estimation. Must be one of "fhlr_2000", "fhlr_2005", "sw_2002", "doz_2011", "doz_2012", or "BM_2014". Default is "fhlr_2000".
\itemize{
  \item \code{"fhlr_2000"}: {Generalized Dynamic-Factor Model (Forni, Hallin, Lippi, and Reichlin, 2000).}
  \item \code{"fhlr_2005"}: {Restricted Generalized Dynamic Factor Model (Forni, Hallin, Lippi, and Reichlin 2005).}
  \item \code{"sw_2002"}: Stock and Watson (2002).
  \item \code{"doz_2011"}: Doz, Giannone, and Reichlin (2011) - relies on the `dfms` package.
  \item \code{"doz_2012"}: Doz, Giannone, and Reichlin (2012) - relies on the `dfms` package.
  \item \code{"BM_2014"}: Banbura and Modugno (2014) - relies on the `dfms` package.
}}
}
\value{
A list containing the results of the DFM estimation. The structure of the list depends on the chosen model.
}
\description{
This function estimates a Dynamic Factor Model (DFM) using various methods.
}
\references{
Forni, M., Hallin, M., Lippi, M., & Reichlin, L. (2000). The generalized dynamic-factormodel: Identification and estimation. The Review of Economics and Statistics.
Forni, M., Hallin, M., Lippi, M., & Reichlin, L. (2005). The generalized dynamicfactor model. Journal of the American Statistical Association, 100(471), 830–840.
Stock, J. H., & Watson, M. W. (2002a). Forecasting using principal componentsfrom a large number of predictors. Journal of the American Statistical Association, 97(460), 1167–1179.
Doz, C., Giannone, D., & Reichlin, L. (2011). A two-step estimator for large ap-proximate dynamic factor models based on kalman filtering. Journal of Econometrics, 164(1), 188-205. 
Doz, C., Giannone, D., & Reichlin, L. (2012, 11). A Quasi–Maximum LikelihoodApproach for Large, Approximate Dynamic Factor Models. The Review of Economics and Statistics, 94(4), 1014-1024.
Banbura, M., & Modugno, M. (2013). Maximum likelihood estimation of large factor model on datasets with arbitrary pattern of missing data. Journal of Applied Econometrics.
}
