% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favar.R
\name{FAVAR}
\alias{FAVAR}
\title{Estimate Factor-Augmented Vector Autoregression (FAVAR)}
\usage{
FAVAR(data, r = 3, n_lags = 13, slow_indices = NULL, direct_indices)
}
\arguments{
\item{data}{A data frame or matrix of data to be analyzed.}

\item{r}{Number of static factors. Default is 3.}

\item{n_lags}{Number of lags in the VAR model. Default is 13.}

\item{slow_indices}{Indices of slow-moving variables. Default is NULL.}

\item{direct_indices}{Indices of directly observed variables.}
}
\value{
A list containing the FAVAR model components:
\itemize{
  \item \code{data_for_var}: Data prepared for VAR.
  \item \code{var}: VAR model.
  \item \code{factors}: Estimated factors.
  \item \code{loadings}: Factor loadings.
}
}
\description{
This function estimates a Factor-Augmented Vector Autoregression (FAVAR) model.
}
\references{
Bernanke, B. S., Boivin, J., & Eliasz, P. (2005, 02). Measuring the Effects of Monetary Policy: A Factor-Augmented Vector Autoregressive (FAVAR) Approach*. The Quarterly Journal of Economics, 120(1), 387-422
}
