% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favar.R
\name{favar_IRF}
\alias{favar_IRF}
\title{Compute Impulse Response Functions (IRFs) for FAVAR}
\usage{
favar_IRF(
  favar,
  shock_variable,
  response_variable = NULL,
  n_ahead = 49,
  n_boot = 500,
  cumulative = FALSE,
  shock = "unit",
  plot = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{favar}{An object of class "FAVAR".}

\item{shock_variable}{The variable to shock.}

\item{response_variable}{The variable to respond. Default is NULL.}

\item{n_ahead}{Number of periods ahead to compute the IRF. Default is 49.}

\item{n_boot}{Number of bootstrap samples. Default is 500.}

\item{cumulative}{Whether to compute cumulative IRFs. Default is FALSE.}

\item{shock}{The type of shock used in the model. "unit" implements a one-unit shock to the schock variable, "sd" implements a shock of one standard deviation. Default is "unit".}

\item{plot}{Whether to plot the IRF. Default is TRUE.}

\item{alpha}{Significance level for the confidence intervals in the plot. Default is 0.05.}
}
\value{
A list containing the IRF and bootstrapped IRF:
\itemize{
  \item \code{irf}: The impulse response function.
  \item \code{boot}: The bootstrapped values of the impulse response function.
}
}
\description{
This function computes the Impulse Response Functions (IRFs) for a Factor-Augmented Vector Autoregression (FAVAR) model.
}
\references{
Bernanke, B. S., Boivin, J., & Eliasz, P. (2005, 02). Measuring the Effects of Monetary Policy: A Factor-Augmented Vector Autoregressive (FAVAR) Approach*. The Quarterly Journal of Economics, 120(1), 387-422
}
