% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_criteria.r
\name{fnets_ic_wrapper}
\alias{fnets_ic_wrapper}
\title{Wrap FNets IC Results}
\usage{
fnets_ic_wrapper(fnets_ic)
}
\arguments{
\item{fnets_ic}{A list or matrix containing the FNets IC results.}
}
\value{
A data frame with the number of factors determined by the FNets IC method.
}
\description{
This function wraps the results from the FNets IC method into a data frame.
}
\references{
Avarucci, M., Cavicchioli, M., Mario, F., & Zaffaroni, P. (2021, 01). The main business cycle shock(s). frequency-band estimation of the number of dynamic factors.
}
