\name{genDFM-package}
\alias{genDFM-package}
\alias{genDFM}
\docType{package}
\title{
  General package to estimate dynamic factor models
}
\description{
genDFM is a general package to estimate dynamic factor models in R.
We provide a set of functions to estimate the model, plot the components or compute 
forecasts. We provide a wide range of information criteria for the choice of the
factors. In addition, the package allows for estimation of Factor Augmented
Vector Autoregressions and computation of the IRFs. 
The core of the package is written in C++ and uses the Armadillo library.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Piotr Radziwanowski, email optional.

Maintainer: Piotr Radziwanowski <p.radziwanowski@student.maastrichtuniversity.nl>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
