% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_criteria.r
\name{onatski_2009_test_r}
\alias{onatski_2009_test_r}
\title{Test k=k_0 vs k_0<k<=k_1}
\usage{
onatski_2009_test_r(
  data,
  omega0 = NULL,
  sequence = NULL,
  k0 = 1,
  k1 = 5,
  test_size = 5
)
}
\arguments{
\item{data}{A matrix of data to be tested.}

\item{omega0}{Optional parameter omega0. Default is NULL.}

\item{sequence}{Optional parameter sequence. Default is NULL.}

\item{k0}{Parameter k0. Default is 1.}

\item{k1}{Parameter k1. Default is 5.}

\item{test_size}{Size of the test. Default is 5.}
}
\value{
The result of the Onatski (2009) test.
}
\description{
This function performs the Onatski (2009) test to compare k=k_0 vs k_0<k<=k_1.
}
\references{
Onatski, A. (2009). Testing hypotheses about the number of factors in large factor models. Econometrica, 77(5), 1447–1479.
}
