% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm.R
\name{plot.DFM}
\alias{plot.DFM}
\title{Plot Dynamic Factor Model (DFM)}
\usage{
\method{plot}{DFM}(x, ...)
}
\arguments{
\item{x}{An object of class "generalized_dfm", "restricted_gdfm", "dfm", or "sw_2002".}

\item{...}{Additional arguments passed to the plot function.}
}
\value{
A plot of the DFM components.
}
\description{
This function plots the components of a Dynamic Factor Model (DFM) based on its class type.
}
