% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favar.R
\name{plot.irf}
\alias{plot.irf}
\title{Plot Impulse Response Functions (IRFs)}
\usage{
\method{plot}{irf}(x, alpha = 0.05)
}
\arguments{
\item{x}{An object of class "irf".}

\item{alpha}{Significance level for the confidence intervals in the plot. Default is 0.05.}
}
\value{
A plot of the IRFs with confidence intervals.
}
\description{
This function plots the Impulse Response Functions (IRFs) along with their confidence intervals.
}
