% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favar.R
\name{plot_IRF}
\alias{plot_IRF}
\title{Plot Impulse Response Functions (IRFs)}
\usage{
plot_IRF(irf, boot_irf, alpha = 0.05)
}
\arguments{
\item{irf}{The impulse response function.}

\item{boot_irf}{The bootstrapped impulse response function.}

\item{alpha}{Significance level for the confidence intervals. Default is 0.05.}
}
\value{
A plot of the IRFs with confidence intervals.
}
\description{
This function plots the Impulse Response Functions (IRFs) along with their confidence intervals.
}
