% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm.R
\name{predict.DFM}
\alias{predict.DFM}
\title{Predict Dynamic Factor Model (DFM)}
\usage{
\method{predict}{DFM}(dfm, h)
}
\arguments{
\item{dfm}{An object of class "generalized_dfm", "restricted_gdfm", "dfm", or "sw_2002".}

\item{h}{The forecast horizon.}
}
\value{
The forecasted values or a message indicating that prediction is not supported for the given class type.
}
\description{
This function makes predictions based on a Dynamic Factor Model (DFM) object.
}
